(maint) abstract #add_env into intermediate class
prior to this change #add_env and its supporting functions were
scattered through the other params like `Command` and `Switch`.  This
made adding a sensitive env type difficult as we would have to further
propagate tests for the new env to all those params.
This change creates an intermediate type on top of `RototillerParam`
which handles `#add_env`