(SLV-124) redact printed command strings
In the previous commits we redact in the param messaging/logging.
But by default, and as an option, you can spit out the command that is
to be run, for easier debugging and transparency.  This may contain
sensitive env var values (which replaced param values).
So we needed a way to print the command string in a safe way, while
still retaining those values to be run silently in the background.  so
here we split out #to_s/to_str with #safe_print.

We had to add an object attribute to track whether a param's value came
from a sensitive env var.